Add support for `kickstart` to `mrack` plugin
